(+semver: feature) Move generated grpc code to pkg
